+++ CHANGELOG.md
@@ -34,3 +34,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Fix errors creating directories when copying Downloads
 - Fix old Minecraft versions sounds not all copying over correctly
 - Fix crashes on boot for ARM devices [#390]
+- Add in tests to forge/fabric servers for network checker tool